🐛 fix: Product 변경으로 인한 캐스팅 타입 수정
🔗 Resolves: #be/feat/198

+++ backend/JiShop/src/main/java/com/jishop/popular/service/PopularCalculationServiceImpl.java
@@ -131,7 +131,7 @@ public class PopularCalculationServiceImpl implements PopularCalculationService 
                 product.getName(),
                 product.getOriginPrice(),
                 product.getDiscountPrice(),
-                productScore.getProduct().getDiscountRate().intValue(),
+                productScore.getProduct().getDiscountRate(),
                 productScore.getTotalOrderCount(),
                 productScore.getReviewRating().doubleValue()
                 // FE 요청으로 리뷰 평점 조정
